requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Login==0.5.0
Flask-WTF==0.15.1
mysql-connector-python==8.0.26
python-dotenv==0.19.0
Flask-Cors==3.0.10
flask-migrate